+++ lib/lowkiq/queue/queue_metrics.rb
@@ -67,7 +67,7 @@ module Lowkiq
 
         return 0 if score.nil?
         return 1 if score == 0 # на случай Actions#perform_all_jobs_now
-        lag = @timestamp.call - score.to_i
+        lag = @timestamp.call - score.to_f
         return 0 if lag < 0
         lag
       end
